// apps/inventory-service/src/stock-ledger/enums.ts
export enum LocationType {
  CENTRAL_WAREHOUSE = 'CENTRAL_WAREHOUSE',
  PHARMACY_COUNTER = 'PHARMACY_COUNTER',
  HOSPITAL_WARD = 'HOSPITAL_WARD',
  LABORATORY = 'LABORATORY',
  SUPPLIER_EXTERNAL = 'SUPPLIER_EXTERNAL',
  PATIENT_OUTBOUND = 'PATIENT_OUTBOUND',
  WASTE_DISPOSAL = 'WASTE_DISPOSAL',
}

export enum TransactionType {
  RECEIVE = 'RECEIVE',
  TRANSFER = 'TRANSFER',
  DISPENSE = 'DISPENSE',
  ADJUSTMENT_ADD = 'ADJUSTMENT_ADD',
  ADJUSTMENT_SUB = 'ADJUSTMENT_SUB',
  EXPIRED_DISCARD = 'EXPIRED_DISCARD',
  RETURN_INBOUND = 'RETURN_INBOUND',
}

export enum ReservationStatusV2 {
  ACTIVE = 'ACTIVE',
  FULFILLED = 'FULFILLED',
  EXPIRED = 'EXPIRED',
  CANCELLED = 'CANCELLED',
}